Start on unnamed single params parsing
Meaning users should be able to call functions that expect a single
call arg to drop the param name from the call, as it can be inferred
when visiting the AST for the function call.

As in:

    my_func(arg = "hello");
    my_func("hello");

are equals in terms of logic.

But

    my_other_func(arg1 = "hello", arg2 = "world");
    my_other_func("hello", arg2 = "world");

are not.

This upcoming change is a QOL addition to make simpler function call
easier to write. This name inferring is optional.